特性: 更新檔案路徑以支援 OneDrive 整合
- 更新本機和 OneDrive 文件的檔案路徑
- 新增條件檢查 OneDrive 文件路徑，若找不到則退出

+++ Windows/copy-profile.ps1
@@ -1,26 +1,38 @@
 # 定義路徑
-$localProfilePath = "$env:USERPROFILE\Documents\PowerShell"
-$oneDriveProfilePath = "$env:OneDrive\文件\PowerShell"
+$localDocumentsPath = "$env:USERPROFILE\Documents"
+$oneDriveDocumentsPath = "$env:OneDrive\文件"
+$localProfilePath = "$localDocumentsPath\PowerShell"
+$oneDriveProfilePath = "$oneDriveDocumentsPath\PowerShell"
 
-# 檢查當前 PowerShell 配置檔案的路徑
-if (Test-Path $localProfilePath) {
-    Write-Output "Local PowerShell profile 路徑存在."
+# 檢查當前使用者的 Documents 路徑
+if (Test-Path $localDocumentsPath) {
+    Write-Output "當前使用者的 Documents 路徑為 Userprofile\Documents."
 
-    $sourceProfile = Join-Path -Path $oneDriveProfilePath -ChildPath "Microsoft.PowerShell_profile.ps1"
-    $destinationProfile = Join-Path -Path $localProfilePath -ChildPath "Microsoft.PowerShell_profile.ps1"
+    # 檢查當前 PowerShell 配置檔案的路徑
+    if (Test-Path $localProfilePath) {
+        Write-Output "Local PowerShell profile 路徑存在."
 
-    if (Test-Path $sourceProfile) {
-        # 複製配置檔案
-        Copy-Item -Path $sourceProfile -Destination $destinationProfile -Force
-        Write-Output "Profile 由 OneDrive 複製至 local PowerShell 資料夾."
+        $sourceProfile = Join-Path -Path $oneDriveProfilePath -ChildPath "Microsoft.PowerShell_profile.ps1"
+        $destinationProfile = Join-Path -Path $localProfilePath -ChildPath "Microsoft.PowerShell_profile.ps1"
 
-        # 加載配置檔案
-        . $PROFILE
-        Write-Output "Profile 加載成功."
+        if (Test-Path $sourceProfile) {
+            # 複製配置檔案
+            Copy-Item -Path $sourceProfile -Destination $destinationProfile -Force
+            Write-Output "Profile 由 OneDrive 複製至 local PowerShell 資料夾."
+
+            # 加載配置檔案
+            . $PROFILE
+            Write-Output "Profile 加載成功."
+        } else {
+            Write-Output "Source profile 未發現於 OneDrive 資料夾."
+        }
     } else {
-        Write-Output "Source profile 未發現於 OneDrive 資料夾."
+        Write-Output "Local PowerShell profile 路徑不存在."
     }
+} elseif (Test-Path $oneDriveDocumentsPath) {
+    Write-Output "當前使用者的 Documents 路徑為 OneDrive\文件, 腳本退出不進行拷貝."
 } else {
-    Write-Output "Local PowerShell profile 路徑不存在."
+    Write-Output "無法找到有效的 Documents 路徑."
+    exit
 }
